requirements.txt
pyyaml
requests
msal
python-dateutil
flask
flask-socketio
flask_session
rich
wheel
gevent-websocket
redis